Move instance-list-group-header to correct folder
