+++ src/DetailsView/reports/components/report-sections/instance-list-group-header.tsx
@@ -3,15 +3,15 @@
 import { kebabCase } from 'lodash';
 import * as React from 'react';
 
-import { GuidanceLinks } from '../../../common/components/guidance-links';
-import { GuidanceTags } from '../../../common/components/guidance-tags';
-import { NewTabLink } from '../../../common/components/new-tab-link';
-import { GetGuidanceTagsFromGuidanceLinks } from '../../../common/get-guidance-tags-from-guidance-links';
-import { NamedSFC } from '../../../common/react/named-sfc';
-import { RuleResult } from '../../../scanner/iruleresults';
-import { OutcomeChip } from './outcome-chip';
-import { outcomeTypeSemantics } from './outcome-type';
-import { InstanceOutcomeType } from './report-sections/outcome-summary-bar';
+import { GuidanceLinks } from '../../../../common/components/guidance-links';
+import { GuidanceTags } from '../../../../common/components/guidance-tags';
+import { NewTabLink } from '../../../../common/components/new-tab-link';
+import { GetGuidanceTagsFromGuidanceLinks } from '../../../../common/get-guidance-tags-from-guidance-links';
+import { NamedSFC } from '../../../../common/react/named-sfc';
+import { RuleResult } from '../../../../scanner/iruleresults';
+import { OutcomeChip } from '../outcome-chip';
+import { outcomeTypeSemantics } from '../outcome-type';
+import { InstanceOutcomeType } from './outcome-summary-bar';
 
 export type InstanceListGroupHeaderDeps = {
     getGuidanceTagsFromGuidanceLinks: GetGuidanceTagsFromGuidanceLinks;
