+++ src/DetailsView/reports/components/report-sections/rule-detail.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 
 import { NamedSFC } from '../../../../common/react/named-sfc';
 import { RuleResult } from '../../../../scanner/iruleresults';
-import { InstanceListGroupHeader, InstanceListGroupHeaderDeps } from '../instance-list-group-header';
+import { InstanceListGroupHeader, InstanceListGroupHeaderDeps } from './instance-list-group-header';
 import { InstanceOutcomeType } from './outcome-summary-bar';
 
 export type RuleDetailDeps = InstanceListGroupHeaderDeps;
